requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.54.0
pydantic>=2.12.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
python-multipart>=0.0.18
httpx>=0.28.0

# ✅ NOVAS - Autenticação
python-jose[cryptography]>=3.3.0
passlib[argon2]==1.7.4
argon2-cffi>=23.1.0
email-validator>=2.1.0